First attempt at moving bulletin methods to parent class
